Extract camera-handling code to a separate file
