+++ app3d/src/jsMain/kotlin/diy/lingerie/web_tool_3d/cameraUtils.kt
@@ -1,0 +1,44 @@
+package diy.lingerie.web_tool_3d
+
+import dev.toolkt.dom.pure.PureSize
+import dev.toolkt.geometry.Vector3
+import dev.toolkt.reactive.cell.Cell
+import three.THREE
+import kotlin.math.PI
+
+fun createRotatoryCamera(
+    height: Double,
+    distance: Double,
+    size: Cell<PureSize>,
+    rotation: Cell<Double>,
+): Pair<THREE.Group, THREE.PerspectiveCamera> {
+    val camera = createReactivePerspectiveCamera(
+        position = Cell.of(
+            Vector3(
+                x = 0.0,
+                y = -distance,
+                z = 0.0,
+            ),
+        ),
+        rotation = Cell.of(
+            THREE.Euler(
+                x = PI / 2,
+            ),
+        ),
+        size = size,
+        fov = 75.0,
+        near = 0.1,
+        far = 1000.0,
+    )
+
+    val group = createReactiveGroup(
+        position = Cell.of(Vector3(x = 0.0, y = 0.0, z = height)),
+        rotation = rotation.map { THREE.Euler(z = it) },
+        children = listOf(camera),
+    )
+
+    return Pair(
+        group,
+        camera,
+    )
+}
